requirements.txt
beautifulsoup4==4.14.3
modules==1.0.0
ollama==0.6.1
pandas==3.0.1
python-dotenv==1.2.1
Requests==2.32.5
streamlit==1.52.2
torch==2.10.0
